first pass as jspm, with node.js file for verifying
