+++ index.html
@@ -1,0 +1,6 @@
+<!doctype html>
+<script src="jspm_packages/system.js"></script>
+<script src="config.js"></script>
+<script>
+System.import('src/app.js');
+</script>
