+++ karma.conf.js
@@ -1,12 +1,15 @@
 // karma.conf.js
 module.exports = function(config) {
 	config.set({
-		basePath: './src',
-		frameworks: ['jasmine'],
+		frameworks: ['jspm', 'jasmine'],
 		browsers: ['PhantomJS2'],
-		files: [
-			'../node_modules/lodash/index.js',
-			'**/*.js'
+		jspm: {
+			loadFiles: [
+				'src/**/*.spec.js'
+			],
+			serveFiles: [
+				'src/**/!(*.spec).js'
 			]
+		}
 	});
 };
