+++ package.json
@@ -8,10 +8,24 @@
   },
   "dependencies": {},
   "devDependencies": {
+    "express": "^4.13.3",
     "jasmine-core": "^2.3.4",
+    "jspm": "^0.16.15",
     "karma": "^0.13.15",
+    "karma-babel-preprocessor": "^6.0.1",
     "karma-jasmine": "^0.3.6",
+    "karma-jspm": "^2.0.2",
     "karma-phantomjs2-launcher": "^0.3.2",
     "lodash": "^3.10.1"
+  },
+  "jspm": {
+    "dependencies": {
+      "lodash": "npm:lodash@^3.10.1"
+    },
+    "devDependencies": {
+      "babel": "npm:babel-core@^5.8.24",
+      "babel-runtime": "npm:babel-runtime@^5.8.24",
+      "core-js": "npm:core-js@^1.1.4"
+    }
   }
 }
