+++ src/app.spec.js
@@ -1,155 +1,6 @@
-// import Startup
-// import assets
-// import Game
-// import Funding
+'use strict';
 
-// TOOD: swithc to Immutable
-var Immutable = {
-	fromJS: function (obj) {
-		return _.extend({}, obj, {
-			get: function (attr) {
-				return this.attrs[attr];
-			},
-			set: function (attr, val) {
-				this.attrs[attr] = val;
-				return this.attrs[attr];
-			}
-		});
-	}
-};
-
-var FUNDING_ROUNDS = ['Bootstrap', 'Seed', 'A', 'B', 'C', 'D'];
-
-/*
-example startup:
-*/
-// TODO: move to own file
-var Startup = function (options) {
-
-	var startup = {};
-
-	startup.attrs = Immutable.fromJS(_.extend({
-		funding: 0,
-		assets: [],
-		currentTurn: 1
-	}, options));
-	startup.get = startup.attrs.get;
-	startup.set = startup.attrs.set;
-
-	startup.setValuation = function (funding) {
-		this.set('valuation', funding * 4);
-	};
-
-	startup.purchaseAsset = function (asset) {
-		this.addAsset(asset);
-		this.decreaseFunding(asset.get('purchaseCost'));
-	};
-
-	startup.addAsset = function (asset) {
-		this.set('assets', this.get('assets').push(asset));
-	};
-
-	startup.addFunding = function (funding) {
-		this.increaseFunding(funding.get('amount'));
-		if (funding.get('type') !== 'bridge') {
-			this.increaseValuation(funding.get('amount') * 4);
-			this.increaseRound(funding.get('round'));
-		}
-	};
-
-	startup.increaseRound = function (round) {
-		var fundingRoundIndex = FUNDING_ROUNDS.indexOf(round);
-		var nextRound;
-
-		if (fundingRoundIndex === FUNDING_ROUNDS.length) {
-			nextRound = FUNDING_ROUNDS[fundingRoundIndex];
-		} else if (fundingRoundIndex !== -1) {
-			nextRound = FUNDING_ROUNDS[fundingRoundIndex + 1];
-		} else {
-			nextRound = this.get('round');
-		}
-
-		this.set('round', nextRound);
-	}
-
-	startup.increaseFunding = function (amount) {
-		this.set('funding', this.get('funding') + amount);
-	};
-
-	startup.decreaseFunding = function (amount) {
-		this.set('funding', this.get('funding') - amount);
-	};
-
-	startup.increaseValuation = function (amount) {
-		this.set('valuation', this.get('valuation') + amount);
-	};
-
-	startup.manageAsset = function (assetId) {
-		var asset = this.getAssetById(id);
-		// TODO: emit error if not found
-		if (asset) {
-			this.increaseValuation(asset.get('valuationPerTurn'));
-		}
-	};
-
-	startup.getAssetById = function (id) {
-		return _.find(this.get('assets'), function (asset) {
-			return asset.get('id') === assetId;
-		});
-	};
-
-	startup.nextTurn = function () {
-		this.get('assets').forEach(function (asset) {
-			this.startup.decreaseFunding(asset.get('perTurnCost'));
-		});
-		this.increaseTurn(1);
-	}
-
-	startup.increaseTurn = function (num) {
-		this.set('currentTurn', this.get('currentTurn') + num);
-	}
-
-	if (!startup.get('valuation')) {
-		startup.setValuation(startup.get('funding'));
-	}
-
-	return startup;
-};
-
-// TODO: move to separate file
-var Funding = function (options) {
-
-	var funding = {};
-
-	funding.attrs = Immutable.fromJS(_.extend({
-		amount: 0,
-		round: FUNDING_ROUNDS[1]
-	}, options));
-	funding.get = funding.attrs.get;
-	funding.set = funding.attrs.set;
-	return funding;
-};
-
-/*
-example asset:
-TODO: move to own file
-*/
-var Asset = function (options) {
-	var asset = {};
-	
-	asset.attrs = Immutable.fromJS({
-		purchaseCost: 1000,
-		perTurnCost: 30,
-		valuationPerTurn: 30*4
-	});
-	asset.get = asset.attrs.get;
-	asset.set = asset.attrs.set;
-	return asset;
-}
-
-var assets = [
-	// TODO: create assets
-];
+import { Startup, Funding, assets, FUNDING_ROUNDS } from './app';
 
 describe('startup init', function () {
